Fix pipeline bugs and complete V3.0 end-to-end validation
- Data: direction-aware reader fix (response fc was always 0), normalizer
  stats/log_bounds persisted, inter_arrival routed to empirical sampling
- Assembler: full 12-function-code builders, response echoes request
  fc/unit_id, per-(fc,direction) expected_fields metadata
- Checker: final timeout sweep, missing-field reporting, exception-fc check,
  MBAP PDU truncation, descriptor parse fixes
- Training: bs=256 config, StubSampler distributions persisted to checkpoint
- Eval: KS/JSD metrics recorded (Mean KS=0.093, protocol layer zero errors)

+++ diffusion/models/masked_diffusion.py
@@ -190,7 +190,7 @@ class MaskedDiffusion(nn.Module):
             cond = s_cond
 
         # Reverse steps: map k_rev ∈ [0, num_unmask_steps) → actual K steps
-        step_indices = torch.linspace(self.K - 1, 0, num_unmask_steps, dtype=torch.long)
+        step_indices = torch.linspace(self.K - 1, 0, num_unmask_steps, dtype=torch.float).long()
 
         for step_idx in range(num_unmask_steps):
             k = step_indices[step_idx]
